requirements.txt
# requirements.txt
Flask==3.0.3
prometheus_client==0.19.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
requests~=2.31.0
grpcio>=1.51.0
protobuf>=4.21.0
google-auth>=2.22.0
